requirements.txt
python-dotenv==1.0.1
Flask==2.2.2
Werkzeug==2.2.2
Flask-Babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
SQLAlchemy==1.4.52
alembic==1.13.2
Mako==1.3.5
WTForms==3.1.2
email-validator==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
blinker==1.8.2
click==8.1.7
Babel==2.15.0
PyJWT==2.8.0
psycopg2-binary==2.9.10

numpy==1.26.4
scipy==1.11.4
pandas==2.2.3
matplotlib==3.8.4
scikit-learn==1.4.2
seaborn==0.13.2
Pillow==10.4.0
plotly==5.24.1
contourpy==1.2.1
fonttools==4.53.1
kiwisolver==1.4.5
pyparsing==3.1.2
cycler==0.12.1

torch==2.5.1

jupyterlab==4.2.5
notebook==7.2.2
jupyter_server==2.14.2
jupyterlab_server==2.27.3
jupyter_client==8.6.3
jupyter_core==5.7.2
ipykernel==6.29.5
ipython==8.30.0
jupyterlab-git==0.50.1
nbconvert==7.16.4
nbformat==5.10.4
nbclient==0.10.0
nbdime==4.0.2
jupyter-events==0.12.0
jupyter-server-mathjax==0.2.6
jupyterlab-pygments==0.3.0
jupyter_server_terminals==0.5.3
mistune==3.0.2
notebook_shim==0.2.4
tornado==6.4.2

djlint==1.36.4
jsonschema==4.23.0
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.7
PyYAML==6.0.2
beautifulsoup4==4.12.3
bleach==6.1.0
GitPython==3.1.43
gitdb==4.0.11
greenlet==3.0.3
joblib==1.4.2
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
packaging==24.1
prometheus-client==0.20.0
psutil==6.0.0
pyzmq==26.2.0
regex==2023.12.25
six==1.16.0
python-dateutil==2.9.0
pytz==2024.1
traitlets==5.14.3
nest-asyncio==1.6.0
terminado==0.18.1
Send2Trash==1.8.3
websocket-client==1.8.0
defusedxml==0.7.1
tinycss2==1.3.0
webencodings==0.5.1
fqdn==1.5.1
isoduration==20.11.0
anyio==4.4.0
argon2-cffi==23.1.0
cffi==1.17.0
attrs==24.2.0
json5==0.9.25
jsonpointer==3.0.0
dnspython==2.6.1
dominate==2.9.1
visitor==0.1.3
EditorConfig==0.12.4
cssbeautifier==1.14.11
jsbeautifier==1.14.11
html-tag-names==0.1.2
html-void-elements==0.1.0
pathspec==0.12.1
colorama==0.4.6
fastjsonschema==2.20.0
importlib-metadata==8.2.0
zipp==3.20.0
stack-data==0.6.3
executing==2.0.1
asttokens==2.4.1
pure-eval==0.2.2
parso==0.8.4
jedi==0.19.1
decorator==5.1.1
matplotlib-inline==0.1.7
prompt-toolkit==3.0.47
wcwidth==0.2.13
pexpect==4.9.0
ptyprocess==0.7.0
Pygments==2.18.0
comm==0.2.2
debugpy==1.8.5
platformdirs==4.2.2
pandocfilters==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
uri-template==1.3.0
webcolors==24.6.0
python-json-logger==2.0.7
sniffio==1.3.1
tenacity==8.5.0
tomli==2.0.1
soupsieve==2.5
smmap==5.0.1
pycparser==2.22
